This patch fixes the feature issue where package is used to redirect a feature to a specific version of the dependency
For example in the case of `diesel` `v1.4.4` the `Cargo.toml` looks like:

```
[dependencies]
uuid = { version = ">=0.2.0, <0.7.0", optional = true, features =
["use_std"] }
uuidv07 = { version = ">=0.7.0, <0.9.0", optional = true, package =
"uuid"}

[features]
...
extras = ["chrono", "serde_json", "uuid", "deprecated-time",
```

If you depend on the feature `uuidv7` which uses a newer `uuid` version
for diesel, `cargo-outdated` will complain that the feature is obsolete
because it is not found in the `features` section but it is links to the
`uuid` package feautre and controls the `uuid` version.

This commit, search through the features, if the feature is not
retained we look for the feature names in the `toml` and also add a
`package` if the dependency has it. Then look through that mapping to
determine if the feature really was obsoleted or if it was a package
derived feature.